totális átírás
