--- Feladatnyilvantarto/src/main/java/Service/LeaderService.java
@@ -1,4 +1,0 @@
-package Service;
-
-public class LeaderService {
-}
