--- Feladatnyilvantarto/src/main/java/Service/RegisterException.java
@@ -1,5 +1,0 @@
-package Service;
-
-public class RegisterException extends RuntimeException
-{
-}
