--- Feladatnyilvantarto/src/main/java/Service/UserService.java
@@ -1,4 +1,0 @@
-package Service;
-
-public class UserService {
-}
