--- Feladatnyilvantarto/src/main/java/Service/WorkerService.java
@@ -1,4 +1,0 @@
-package Service;
-
-public class WorkerService {
-}
