+++ Feladatnyilvantarto/src/main/java/hu/elte/feladatnyilvantarto/FeladatnyilvantartoApplication.java
@@ -1,0 +1,12 @@
+package hu.elte.feladatnyilvantarto;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class FeladatnyilvantartoApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(FeladatnyilvantartoApplication.class, args);
+    }
+}
