+++ Feladatnyilvantarto/src/main/java/hu/elte/feladatnyilvantarto/domain/Credentials.java
@@ -1,6 +1,15 @@
-package Domain;
+package hu.elte.feladatnyilvantarto.domain;
 
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+
+@Entity
 public class Credentials {
+
+    @Id
+    @GeneratedValue
+    private int id;
     private String loginName;
     private String password;
