+++ Feladatnyilvantarto/src/main/java/hu/elte/feladatnyilvantarto/domain/NotificationType.java
@@ -1,4 +1,4 @@
-package Domain;
+package hu.elte.feladatnyilvantarto.domain;
 
 public enum NotificationType {
     GROUP, TICKET, COMMENT
