+++ Feladatnyilvantarto/src/main/java/hu/elte/feladatnyilvantarto/domain/Priority.java
@@ -1,4 +1,4 @@
-package Domain;
+package hu.elte.feladatnyilvantarto.domain;
 
 public enum Priority {
     High,Medium,Low
