+++ Feladatnyilvantarto/src/main/java/hu/elte/feladatnyilvantarto/domain/UserType.java
@@ -1,4 +1,4 @@
-package Domain;
+package hu.elte.feladatnyilvantarto.domain;
 
 public enum UserType {
     Leader,Worker
