+++ Feladatnyilvantarto/src/main/java/hu/elte/feladatnyilvantarto/service/LoginException.java
@@ -1,4 +1,4 @@
-package Service;
+package hu.elte.feladatnyilvantarto.service;
 
 public class LoginException extends RuntimeException
 {
