+++ Feladatnyilvantarto/src/main/java/hu/elte/feladatnyilvantarto/service/RegisterException.java
@@ -1,0 +1,23 @@
+package hu.elte.feladatnyilvantarto.service;
+
+public class RegisterException extends RuntimeException
+{
+    public RegisterException() {
+    }
+
+    public RegisterException(String message) {
+        super(message);
+    }
+
+    public RegisterException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public RegisterException(Throwable cause) {
+        super(cause);
+    }
+
+    public RegisterException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
+        super(message, cause, enableSuppression, writableStackTrace);
+    }
+}
