+++ Feladatnyilvantarto/src/main/java/hu/elte/feladatnyilvantarto/service/UserService.java
@@ -1,0 +1,5 @@
+package hu.elte.feladatnyilvantarto.service;
+
+public class UserService {
+
+}
