+++ Feladatnyilvantarto/src/main/java/hu/elte/feladatnyilvantarto/service/WorkerService.java
@@ -1,0 +1,4 @@
+package hu.elte.feladatnyilvantarto.service;
+
+public class WorkerService {
+}
