+++ Feladatnyilvantarto/src/main/resources/application.properties
@@ -1,0 +1,9 @@
+spring.application.name=Feladatnyilvántartó
+
+spring.datasource.url=jdbc:h2:mem:testdb
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=
+
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+spring.h2.console.enabled=true
